When deleting all formats except ..., do not delete if it leaves a book with no formats

+++ src/calibre/gui2/actions/delete.py
@@ -161,9 +161,12 @@ class DeleteAction(InterfaceAction):
                 continue
             bfmts = set([x.lower() for x in bfmts.split(',')])
             rfmts = bfmts - set(fmts)
-            for fmt in rfmts:
-                self.gui.library_view.model().db.remove_format(id, fmt,
-                        index_is_id=True, notify=False)
+            if bfmts - rfmts:
+                # Do not delete if it will leave the book with no
+                # formats
+                for fmt in rfmts:
+                    self.gui.library_view.model().db.remove_format(id, fmt,
+                            index_is_id=True, notify=False)
         self.gui.library_view.model().refresh_ids(ids)
         self.gui.library_view.model().current_changed(self.gui.library_view.currentIndex(),
                 self.gui.library_view.currentIndex())
